requirements.txt
pandas
pytest
streamlit
google-genai
requests
python-dotenv
scikit-learn
